PR URL: https://msazure.visualstudio.com/DefaultCollection/Liftr/_git/Liftr.Common/pullrequest/3513941

Related work items: #8179930

+++ src/Liftr.Logging.Metrics/IMetricSender.cs
@@ -8,8 +8,8 @@ namespace Microsoft.Liftr.Logging.Metrics
 {
     public interface IMetricSender
     {
-        void Gauge(string metric, double value, Dictionary<string, string> dimension = null);
+        void Gauge(string metric, int value, Dictionary<string, string> dimension = null);
 
-        void Gauge(string mdmNamespace, string metric, double value, Dictionary<string, string> dimension = null);
+        void Gauge(string mdmNamespace, string metric, int value, Dictionary<string, string> dimension = null);
     }
 }
